requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
pytesseract==0.3.10
sqlparse==0.4.2
